fix: compute sigmoid in fp32 in biased_grouped_topk_impl fallback path
All other routing paths (flashinfer, moe_fused_gate, kimi_k2) cast
gating_output to fp32 before computing sigmoid, but the fallback
biased_grouped_topk_impl computed sigmoid in the input dtype (bf16).
This caused routing mismatches for models like GLM-4.7-Flash whose
parameters (n_group=1, topk_group=1, 64 experts) don't hit any fast
path.

+++ python/sglang/srt/layers/moe/topk.py
@@ -645,7 +645,7 @@ def biased_grouped_topk_impl(
 ):
     assert hidden_states.shape[0] == gating_output.shape[0], "Number of tokens mismatch"
 
-    scores = gating_output.sigmoid()
+    scores = gating_output.float().sigmoid().to(gating_output.dtype)
     num_token = scores.shape[0]
     num_experts = scores.shape[1]
     scores_for_choice = scores.view(num_token, -1) + correction_bias.unsqueeze(0)
